shim: add Windows support for manager, I/O, and task
Split platform-specific code into Unix and Windows variants to make
the shim layer compile on Windows.

Manager (internal/shim/manager/):
- Extract shared code (Name, Info, readSpec, groupLabels) into
  manager.go. Move Start/Stop and Unix socket logic to
  manager_unix.go. Add manager_windows.go with named pipe TTRPC
  address (via TTRPC_SOCKET env var), CREATE_NEW_PROCESS_GROUP,
  and a pipe-ready wait loop before returning the address.

I/O streams (internal/shim/task/):
- Extract copyStreams from io.go into io_copystreams_unix.go (FIFO
  via containerd/fifo) and io_copystreams_windows.go (named pipes
  via go-winio). Add "pipe" as an accepted scheme alongside "fifo".
- Add //go:build !windows to io_other.go (defaultScheme conflict).

Networking:
- Remove //go:build !windows from networking_unix.go -> networking.go.
  The code is pure Go (no Unix dependencies) and works on Windows
  as-is. Delete the Windows stub. This enables the shim to parse
  network annotations and call AddNIC on Windows.

Other task fixes:
- ctrnetworking.go: Replace unix.IFNAMSIZ with local constant.
- mount.go: Use path.Join for VM-internal paths (forward slashes).
- mount_other.go: Consolidate with mount_darwin.go (both call
  transformMounts). Add //go:build !linux tag.
