--- internal/shim/task/networking_windows.go
@@ -1,38 +1,0 @@
-/*
-   Copyright The containerd Authors.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package task
-
-import (
-	"context"
-
-	"github.com/containerd/nerdbox/internal/shim/task/bundle"
-	"github.com/containerd/nerdbox/internal/vm"
-)
-
-type networkProvider struct{}
-
-func (p *networksProvider) FromBundle(ctx context.Context, b *bundle.Bundle) error {
-	return nil
-}
-
-func (p *networksProvider) SetupVM(ctx context.Context, vmi vm.Instance) error {
-	return nil
-}
-
-func (p *networksProvider) InitArgs() []string {
-	return nil
-}
